+++ Changelog.md
@@ -6,6 +6,7 @@
 - Added missing methods to the `Zone` interface (have been present in the implementation from v2.5)
 - Added the `getMemberCountryIds()` and the `getMemberProvinceIds()` methods to the Zone class/interface
 - Added the `ofType()` scope to the ZoneMember model
+- Added `ZoneMember::getName()` both to the model and to the interface 
 
 ## 2.6.0
 ##### 2023-03-10